Search repository IT

+++ awbd-be/src/main/java/ro/unibuc/fmi/awbd/common/model/PaginationInformation.java
@@ -7,7 +7,7 @@ import lombok.val;
 @Data
 @Builder
 public class PaginationInformation {
-    private long page;
+    private int page;
     private int pageSize;
     private long pagesTotal;
     private boolean hasNextPage;
